% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\docType{package}
\name{fastverse-package}
\alias{fastverse-package}
\alias{fastverse}
\alias{.matrixStats_replaced}
\title{The \emph{fastverse}: A Suite of High-Performance Packages for Statistical Computing and Data Manipulation in R}
\description{
The \emph{fastverse} is an extensible suite of R packages developed by heterogenous people
and organizations that contribute towards the objectives of:
\enumerate{
\item Speeding up R through heavy use of compiled code (C, C++, Fortran)
\item Enabling more complex statistical and data manipulation operations in R
\item Reducing the number of dependencies required to do advanced computing in R
}
Inspired by the \code{tidyverse} package, the \code{fastverse} package is a flexible package loader and manager 
that installs and by default loads 6 core packages (\code{data.table}, \code{collapse}, \code{matrixStats}, \code{kit}, \code{magrittr} and \code{fst}) that 
are of high quality, provide native C/C++ code, work well together and enable complex statistical computing and data manipulation - with only \code{Rcpp} as an additional dependency. 

The package also harmonises functionality between some of these core packages (see below) and provides additional functionality 
to freely (and permanently) extend the \emph{fastverse} with additional packages, or to reduce the number of packages loaded 
with \code{library(fastverse)}. Further functions help sorting out conflicts between attached packages, and updating packages.
}
\section{Functions in the \emph{fastverse} package}{


Functions to extend or reduce the number of packages in the \emph{fastverse}, either for the session or permanently:

\code{\link[=fastverse_extend]{fastverse_extend()}}\cr
\code{\link[=fastverse_detach]{fastverse_detach()}}

Function to display conflicts between fastverse packages

\code{\link[=fastverse_conflicts]{fastverse_conflicts()}}

Function to update fastverse packages (and dependencies)

\code{\link[=fastverse_update]{fastverse_update()}}

Utilities to retrieve the names of \emph{fastverse} packages (and dependencies), and their update status.

\code{\link[=fastverse_packages]{fastverse_packages()}}\cr
\code{\link[=fastverse_packages]{fastverse_deps()}}\cr
\code{\link[=fastverse_packages]{fastverse_sitrep()}}
}

\section{\emph{fastverse} package options}{

\itemize{
\item Setting \code{option(fastverse_quiet = TRUE)} will disable all automatic messages when calling \code{library(fastvsers)} or \code{\link{fastverse_extend}}
\item \code{option(fastverse_extend = c(...))} can be set before calling \code{library(fastvsers)} to extend the fastverse with some packages for the session. The same can be done with the
\code{\link{fastverse_extend}} function after \code{library(fastvsers)}, which will also populate \code{option("fastverse_extend")}. 
}
}

\section{\emph{fastverse} harmonisations}{

\itemize{
\item There are 2 internal clashes between \code{collapse::funique} and \code{kit::funique}, and between \code{matrixStats::count} and \code{kit::count}.
The \emph{collapse} and \emph{matrixStats} versions take precedence over the \emph{kit} versions. For a comparison see the details section of \code{\link{fastverse_conflicts}}.
\item Quite A number of functions in the \emph{matrixStats} package do not bother about the attributes of objects passed to them. 
The \emph{fastverse} alters a number of these functions, listed in a global variable \code{.matrixStats_replaced}, bestowing them with 
capabilities to preserve matrix dimension names and other attributes for  functions preserving the dimensions of the matrix. 
This is done using very efficient R and C code, so that performance does not suffer. When the \emph{fastverse} package is attached, these
altered function are replaced in the \emph{matrixStats} namespace.      
}
}

