% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastverse_deps}
\alias{fastverse_deps}
\title{List all fastverse dependencies}
\usage{
fastverse_deps(
  recursive = FALSE,
  repos = getOption("repos"),
  include.self = FALSE
)
}
\arguments{
\item{recursive}{logical. \code{TRUE} recursively determines all packages required to operate the current set of \emph{fastverse} packages.
\code{FALSE} will only list the \emph{fastverse} packages and direct dependencies of those packages.}

\item{repos}{the repositories to use to check for updates. Defaults to \code{getOptions("repos")}.}

\item{include.self}{logical. \code{TRUE} also includes the \emph{fastverse} package and checks against CRAN updates.}
}
\value{
A data frame giving the package names, the CRAN and local version, and a logical variable stating whether the local version is behind the CRAN version.
}
\description{
Lists all fastverse dependencies.
}
\seealso{
\code{\link{fastverse_sitrep}}, \code{\link{fastverse}}
}
