% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{fastverse_extend}
\alias{fastverse_extend}
\title{Extend the fastverse with packages}
\usage{
fastverse_extend(..., topics = NULL, install = FALSE, permanent = FALSE)
}
\arguments{
\item{\dots}{comma-separated package names.}

\item{topics}{character. Short-keys to attach groups of related packages:}

\item{install}{logical. Install packages not available?}

\item{permanent}{logical. Should extensions be saved and included when \code{library(fastverse)} is called next time. This is implemented via a config file saved to the package directory.}
}
\description{
Loads additional packages as part of the fastverse.
}
\examples{
fastverse_extend(xts, roll)
}
