% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastverse_packages}
\alias{fastverse_packages}
\title{\emph{Fastverse} Options}
\usage{
fastverse_packages(extended = TRUE, include.self = TRUE)
}
\arguments{
\item{extended}{logical. \code{TRUE} return all packages currently loaded with \code{library(fastverse)}. \code{FALSE} only returns the core fastverse packages.
For a list of suggested packages to extend \emph{fastverse} see \code{\link{fastverse_extend}} or the README page.}

\item{include.self}{logical. Include the \emph{fastverse} package in the list?}
}
\description{
Setting \code{options(fastverse_quiet = TRUE)} will not print any messages to the console when calling \code{library(fastverse)}.
List all packages in the fastverse
}
\examples{
fastverse_packages()
}
