% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastverse_reset}
\alias{fastverse_reset}
\title{Reset the fastverse to defaults}
\usage{
fastverse_reset(options = TRUE)
}
\arguments{
\item{options}{logical. \code{TRUE} also clears all \emph{fastverse} options.}
}
\value{
This function returns nothing. Courtesy to CRAN maintainers, and because my name is not Hadley Wickham - who ostensibly is able to publish an entire \emph{tidyverse} package without these redundant statements.
}
\description{
Calling this function will remove global configuration files and (default) clear all package options. 
Attached packages will not be detached, and configuration files for projects (as discussed in the vignette) will not be removed.
}
\seealso{
\code{\link{fastverse_extend}}, \code{\link{fastverse}}
}
