% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{fastversechild_conflicts}
\alias{fastversechild_conflicts}
\title{Conflicts between the fastversechild and other packages}
\usage{
fastversechild_conflicts(pkg = fastversechild_packages())
}
\arguments{
\item{pkg}{character. A vector of packages to check conflicts for. The default is all \emph{fastversechild} packages.}
}
\value{
An object of class 'fastversechild_conflicts': A named list of character vectors where the names are the conflicted objects, and the 
content are the names of the package namespaces containing the object, in the order they appear on the \code{\link{search}} path.
}
\description{
This function lists all the conflicts among \emph{fastversechild} packages and between \emph{fastversechild} packages and other attached packages.
It can also be used to check conflicts for any other attached packages.
}
\examples{
# Check conflicts between fastversechild packages and all attached packages
fastversechild_conflicts()

# Check conflicts among all attached packages
fastversechild_conflicts(rm_stub(search()[-1], "package:"))
}
\seealso{
\code{\link{fastversechild}}
}
