% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{fastversechild_detach}
\alias{fastversechild_detach}
\title{Detach fastversechild packages}
\usage{
fastversechild_detach(
  ...,
  unload = FALSE,
  force = FALSE,
  include.self = TRUE,
  session = FALSE
)
}
\arguments{
\item{\dots}{comma-separated package names, quoted or unquoted, or vectors of package names. If left empty, all packages returned by \code{\link{fastversechild_packages}} are detached.}

\item{unload}{logical. \code{TRUE} also unloads the packages using \code{\link[=detach]{detach(name, unload = TRUE)}}.}

\item{force}{logical. should a \emph{fastversechild} package be detached / unloaded even though other attached packages depend on it?}

\item{include.self}{logical. \code{TRUE} also includes the \code{fastversechild} package - only applicable if \code{\dots} is left empty.}

\item{session}{logical. \code{TRUE} also removes the packages from \code{options("fastversechild.extend")}, so they will not be attached again with \code{library(fastversechild)} in the current session. If \code{\dots} is left empty and \code{include.self = TRUE}, this will clear \bold{all} \emph{fastversechild} options set for the session.}
}
\description{
Detaches \emph{fastversechild} packages, removing them from the \code{\link{search}} path.
}
\seealso{
\code{\link{fastversechild_extend}}, \code{\link{fastversechild}}
}
