% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{fastversechild_extend}
\alias{fastversechild_extend}
\title{Extend the fastversechild}
\usage{
fastversechild_extend(
  ...,
  install = FALSE,
  check.conflicts = !isTRUE(getOption("fastversechild.quiet"))
)
}
\arguments{
\item{\dots}{comma-separated package names, quoted or unquoted, or vectors of package names.}

\item{install}{logical. Install packages not available?}

\item{check.conflicts}{logical. Should conflicts between extension packages and attached packages be checked?}
}
\description{
Loads additional packages as part of the \emph{fastversechild}.
}
\details{
When the \emph{fastversechild} is extended calling \code{fastversechild_extend(...)}, the packages that are not attached are attached, but conflicts are checked for all specified packages. 
An \code{options("fastversechild.extend")} is set which stores these extension packages, regardless of whether they were already attached or not. When calling 
\code{\link{fastversechild_packages}}, \code{\link{fastversechild_deps}}, \code{\link{fastversechild_conflicts}}, \code{\link{fastversechild_update}}, \code{\link{fastversechild_sitrep}} or \code{\link{fastversechild_detach}}, these packages are included as part of the \emph{fastversechild}. 
To extend the \emph{fastversechild} for the current session when it is not yet loaded, users can also set \code{options(fastversechild.extend = c(...))}, where \code{c(...)}
is a character vector of package names, before calling \code{library(fastversechild)}.
}
\seealso{
\code{\link{fastversechild_detach}}, \code{\link{fastversechild}}
}
