% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastversechild_install}
\alias{fastversechild_install}
\title{Install (missing) fastversechild packages}
\usage{
fastversechild_install(..., only.missing = TRUE, install = TRUE)
}
\arguments{
\item{\dots}{comma-separated package names, quoted or unquoted, or vectors of package names. If left empty, all packages returned by \code{\link{fastversechild_packages}} are checked.}

\item{only.missing}{logical. \code{TRUE} only installs packages that are unavailable. \code{FALSE} installs all packages, even if they are available.}

\item{install}{logical. \code{TRUE} will proceed to install packages, whereas \code{FALSE} (recommended) will print the installation command asking you to run it in a clean R session.}
}
\value{
\code{fastversechild_install} returns \code{NULL} invisibly.
}
\description{
This function (by default) checks if any \emph{fastversechild} package is missing and installs the missing package(s).
}
\details{
This function is useful to call before \code{library(fastversechild)}, especially in a project when a \code{.fastversechild} configuration file is used to load a custom set of packages not natively installed by the \code{fastversechild} package (see \href{https://fastverse.github.io/fastverse/articles/fastverse_intro.html}{fastverse vignette}). 
The idea behind this is that you may customize the \emph{fastversechild} for the project, then revisit the project after a long while, and some packages listed in your config file may be missing. 
You can guard against missing packages by running \code{fastversechild::fastversechild_install()} before \code{library(fastversechild)}, to ensure that all \emph{fastversechild} packages are available before loading the \emph{fastversechild} and executing any code.
}
\seealso{
\code{\link{fastversechild_update}}, \code{\link{fastversechild}}
}
