% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastversechild_packages}
\alias{fastversechild_packages}
\title{List all packages in the fastversechild}
\usage{
fastversechild_packages(extensions = TRUE, include.self = TRUE)
}
\arguments{
\item{extensions}{logical. \code{TRUE} appends the set of core packages with all packages found in \code{options("fastversechild.extend")}.}

\item{include.self}{logical. Include the \emph{fastversechild} package in the list?}
}
\value{
A character vector of package names.
}
\description{
Core packages are first fetched from a project-level configuration file (named \code{.fastversechild}, if found), otherwise the standard set of core packages is returned. 
In addition, if \code{extensions = TRUE}, any packages used to extend the \emph{fastversechild} for the current 
session are also returned.
}
\examples{
fastversechild_packages()
}
\seealso{
\code{\link{fastversechild_extend}}, \code{\link{fastversechild}}
}
