% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastversechild_sitrep}
\alias{fastversechild_sitrep}
\title{Get a situation report on the fastversechild}
\usage{
fastversechild_sitrep(...)
}
\arguments{
\item{\dots}{arguments other than \code{pkg} passed to \code{\link{fastversechild_deps}}.}
}
\value{
\code{fastversechild_sitrep} returns \code{NULL} invisibly.
}
\description{
This function gives a quick overview of the version of R and all 
\emph{fastversechild} packages (including availability updates for packages) and indicates 
whether any global or project-level configuration files are used.
}
\seealso{
\code{\link{fastversechild_deps}}, \code{\link{fastversechild}}
}
