% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastversechild_update}
\alias{fastversechild_update}
\title{Update fastversechild packages}
\usage{
fastversechild_update(..., install = FALSE)
}
\arguments{
\item{\dots}{arguments passed to \code{\link{fastversechild_deps}}.}

\item{install}{logical. \code{TRUE} will proceed to install outdated packages, whereas \code{FALSE} (recommended) will print the installation command asking you to run it in a clean R session.}
}
\value{
\code{fastversechild_update} returns \code{NULL} invisibly.
}
\description{
This will check all \emph{fastversechild} packages (and their
dependencies) for updates and (optionally) install those updates.
}
\seealso{
\code{\link{fastversechild_deps}}, \code{\link{fastversechild}}
}
