% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_attached}
\alias{is_attached}
\alias{is_installed}
\title{Utilities}
\usage{
is_attached(x)

is_installed(x)
}
\arguments{
\item{x}{character. A vector of package names.}
}
\description{
Checks if packages are installed or attached.
}
