% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{fastverse_conflicts}
\alias{fastverse_conflicts}
\title{Conflicts between the fastverse and other packages}
\usage{
fastverse_conflicts(pck = fastverse_packages(include.self = FALSE))
}
\arguments{
\item{pck}{character. A string of packages to check conflicts for. The default is all fastverse packages.}
}
\description{
This function lists all the conflicts between packages in the fastverse
and other packages that you have loaded.
}
\details{
There is an internal conflict between \code{collapse::funique} and \code{kit::funique}. 
If both packages are unloaded, \code{collapse} is loaded before \code{kit}. In general the 
\code{collapse} version is often faster on data frames whereas the \code{kit} version is generally
faster for vectors and also supports matrices.
}
\examples{
fastverse_conflicts()
}
