% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastverse_deps}
\alias{fastverse_deps}
\title{List all \emph{fastverse} dependencies}
\usage{
fastverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{if \code{TRUE}, will also list all dependencies of
\emph{fastverse} packages.}

\item{repos}{the repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all \emph{fastverse} dependencies
}
