% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{fastverse_detach}
\alias{fastverse_detach}
\title{Detach fastverse packages}
\usage{
fastverse_detach(unload = TRUE, include.self = TRUE, force = FALSE)
}
\arguments{
\item{unload}{logical. \code{TRUE} also unloads the package using \code{\link[=detach]{detach(name, unload = TRUE)}}.}

\item{include.self}{logical. \code{TRUE} also detaches the fastverse package.}

\item{force}{logical. should a package be detached even though other attached packages depend on it?}
}
\description{
Detaches fastverse packages (removing them from the \code{\link{search}} path).
}
\examples{
fastverse_detach()
}
