% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastverse_packages}
\alias{fastverse_packages}
\title{List all packages in the fastverse}
\usage{
fastverse_packages(extended = TRUE, include.self = TRUE)
}
\arguments{
\item{extended}{logical. \code{TRUE} return all packages currently loaded with \code{library(fastverse)}. \code{FALSE} only returns the core fastverse packages.
For a list of suggested packages to add to your fastverse see the README page.}

\item{include.self}{logical. Include the fastverse package in the list?}
}
\description{
List all packages in the fastverse
}
\examples{
fastverse_packages()
}
