% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastverse_packages}
\alias{fastverse_packages}
\title{List all packages in the fastverse}
\usage{
fastverse_packages(extensions = TRUE, include.self = TRUE)
}
\arguments{
\item{extensions}{logical. \code{TRUE} appends the set of core packages with all packages found in \code{options("fastverse.extend")}.}

\item{include.self}{logical. Include the \emph{fastverse} package in the list?}
}
\value{
A character vector of package names.
}
\description{
Core packages are first fetched from a project-level configuration file (if found), else from a global configuration file (if found), 
otherwise the standard set of core packages is returned. In addition, if \code{extensions = TRUE}, any packages used to extend the \emph{fastverse} for the current 
session (fetched from \code{getOption("fastverse.extend")}) are also returned.
}
\examples{
fastverse_packages()
}
\seealso{
\code{\link{fastverse_extend}}, \code{\link{fastverse}}
}
