% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastverse_packages}
\alias{fastverse_packages}
\title{List all packages in the fastverse}
\usage{
fastverse_packages(extended = TRUE, include.self = TRUE)
}
\arguments{
\item{extended}{logical. Include extension packages to the core fastverse.}

\item{include.self}{logical. Include the fastverse package in the list?}
}
\description{
List all packages in the fastverse
}
\examples{
fastverse_packages()
}
