% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastverse_reset}
\alias{fastverse_reset}
\title{Reset the fastverse to defaults}
\usage{
fastverse_reset(options = TRUE)
}
\arguments{
\item{options}{logical. \code{TRUE} also clears all \emph{fastverse} options.}
}
\description{
Calling this function will remove global configuration \emph{fastverse} files and (default) clear all package options. 
Packages loaded will not be unloaded, and configuration files for projects (as discussed in the vignette) will not be removed.
}
\seealso{
\code{\link{fastverse_extend}}, \code{\link{fastverse}}
}
