% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastverse_sitrep}
\alias{fastverse_sitrep}
\title{Get a situation report on the fastverse}
\usage{
fastverse_sitrep(...)
}
\arguments{
\item{\dots}{arguments other than \code{pkg} passed to \code{\link{fastverse_deps}}.}
}
\description{
This function gives a quick overview of the version of R and all 
\emph{fastverse} packages (including availability updates for packages) and indicates 
whether any global or project-level configuration files are used
(as described in more detail the vignette).
}
\seealso{
\code{\link{fastverse_deps}}, \code{\link{fastverse}}
}
