% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastverse_update}
\alias{fastverse_update}
\title{Update fastverse packages}
\usage{
fastverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{if \code{TRUE}, will also list all dependencies of
\emph{fastverse} packages.}

\item{repos}{the repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all fastverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
fastverse_update()
}
}
