% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastverse_update}
\alias{fastverse_update}
\title{Update fastverse packages}
\usage{
fastverse_update(check.deps = TRUE, ...)
}
\arguments{
\item{check.deps}{logical. \code{TRUE} also checks for updates in dependencies of \emph{fastverse} packages.}

\item{\dots}{arguments passed to \code{\link{fastverse_deps}}.}
}
\description{
This will check all fastverse packages (and their
dependencies) for updates and print an install command.
}
\seealso{
\code{\link{fastverse_deps}}, \code{\link{fastverse}}
}
