% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fastverse_update}
\alias{fastverse_update}
\title{Update fastverse packages}
\usage{
fastverse_update(...)
}
\arguments{
\item{\dots}{arguments passed to \code{\link{fastverse_deps}}.}
}
\description{
This will check all \emph{fastverse} packages (and their
dependencies) for updates and print a command to install those updates.
}
\seealso{
\code{\link{fastverse_deps}}, \code{\link{fastverse}}
}
